Employee delete endpoint access granted to admin users only

+++ src/main/java/com/emp/employee_management_system/Controllers/EmployeeController.java
@@ -1,8 +1,12 @@
 package com.emp.employee_management_system.Controllers;
 
 import com.emp.employee_management_system.Dto.EmployeeDto;
+import com.emp.employee_management_system.Jwt.JwtService;
 import com.emp.employee_management_system.Services.EmployeeService;
+import com.emp.employee_management_system.Utils.Response.ResponseHandler;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -13,6 +17,9 @@ public class EmployeeController {
 
     @Autowired
     private EmployeeService employeeService;
+
+    @Autowired
+    private JwtService jwtService;
 
     @GetMapping
     ResponseEntity<Object> getAllEmployees() {
@@ -30,17 +37,36 @@ public class EmployeeController {
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<Object> deleteEmployee(@PathVariable Long id) {
-        return employeeService.deleteEmployee(id);
+    ResponseEntity<Object> deleteEmployee(@PathVariable Long id,@RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader) {
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            String jwtToken = authorizationHeader.substring(7);
+
+            String roles = jwtService.getRoles(jwtToken);
+            if (roles.contains("ADMIN")) {
+                return employeeService.deleteEmployee(id);
+            } else {
+                return ResponseHandler.generateResponse(
+                        "Access denied !",
+                        HttpStatus.FORBIDDEN,
+                        null
+                );
+            }
+        } else {
+            return ResponseHandler.generateResponse(
+                    "Invalid token !",
+                    HttpStatus.UNAUTHORIZED,
+                    null
+            );
+        }
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<Object> deleteEmployee(@PathVariable Long id, @RequestBody EmployeeDto employeeDto) {
+    ResponseEntity<Object> UpdateEmployee(@PathVariable Long id, @RequestBody EmployeeDto employeeDto) {
         return employeeService.updateEmployee(id,employeeDto);
     }
 
     @PostMapping("/uploadProfilePic/{id}")
-    ResponseEntity<Object> deleteEmployee(@PathVariable Long id, @RequestParam("profileImage") MultipartFile profileImage) {
+    ResponseEntity<Object> UploadProfilePic(@PathVariable Long id, @RequestParam("profileImage") MultipartFile profileImage) {
         return employeeService.uploadProfilePicture(id,profileImage);
     }
